requirements.txt
transformers #==4.40.0
datasets #==2.19.0
sentencepiece #==0.2.0
sacrebleu #==2.4.0
torch #==2.2.0
accelerate #==0.29.0
evaluate #==0.4.1
numpy #==1.26.4
pandas #==2.2.1
tqdm #==4.66.2
peft #==0.10.0
wandb #==0.16.6
langdetect #==1.0.9
python-dotenv #==1.0.1
pytest #==7.3.1
gradio #==3.34.0
google-cloud-storage #>=2.10.0